Fix flaky/erratic tests in Pleroma.Config.TransferTaskTest
There were async calls happening, so they weren't always finished when assert happened.

I also fixed some bugs in the erratic tests that were introduced when removing :shout.
:shout is a key where restart is needed, and was changed in the test to use :rate_limit
But there was a bug in the syntax that didn't get caught because the test was tagged as
erratic and therefor didn't fail. Here I fixed it.