Fix config path lookup
While I was attempting to set up an Akkoma instance using the OTP build,
I noticed that it was falling back to using `/etc/pleroma` instead of
`/etc/akkoma` for the config path.

As the `config.exs` file doesn't exist prior to running
`pleroma_ctl instance gen`, it tries to write to
`/etc/pleroma/config.exs` but the directory doesn't exist when following
the OTP install guide, so the config generation fails.

This patch should hopefully fix that.