Add translation module for Argos Translate
Argos Translate is a Python module for translation and can be used as a command line tool.
This is also the engine for LibreTranslate, for which we already have a module.
Here we can use the engine irectly from our server without doing requests to a third party or
having to install our own LibreTranslate webservice.

One thing that's currently still missing from ArgosTranslate is auto-detection of languages.