Allow stripping of html for Argos Translate module
Argos Translate doesn't properly translate HTML.
Libre Translate uses an extra module for that.
Here we strip the HTML by default before translating and then add rudimentary HTML back so it displays properly.

Stripping HTML is a default-on option, just in case argos adds support for html in the language models later on.
That way admins will be able to overwrite the setting and not strip any more.