Add docker-compose.yml to gitignore
The docker-compose.yml file is likely to be edited quite extensively by
admins when setting up an instance. To make it easier to do upgrades
down the line, it should be ignored by git.

Instead we should provide the existing docker-compose.yml as a sample
config and update the docs to have the user copy it over to the root
directory to edit it themselves.